pyproject.toml
[project]
name = "silkroad-ecovision"
version = "0.1.0"
description = "CSIRO Image2Biomass reproduction pipeline with uv-managed local development."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "albumentations>=1.4.18",
    "numpy>=2.1.2",
    "opencv-python-headless>=4.10.0.84",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "pyarrow>=18.1.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.5.2",
    "timm>=1.0.11",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.66.6",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.3",
    "ruff>=0.7.2",
]

[project.scripts]
csiro-biomass = "csiro_biomass.cli.main:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/csiro_biomass"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
